removed megatron, moved smooth cross entropy
