+++ nemo/collections/common/tokenizers/tokenizer_spec.py
@@ -21,6 +21,9 @@ __all__ = ['TokenizerSpec']
 
 
 class TokenizerSpec(ABC):
+    """
+    Inherit this class to implement a new tokenizer.
+    """
     @abstractmethod
     def text_to_tokens(self, text):
         pass
