+++ nemo/collections/nlp/models/lm_model.py
@@ -13,11 +13,9 @@
 # limitations under the License.
 
 import os
-from typing import Dict, List, Optional
+from typing import Dict, Optional
 
 import torch
-import torch.nn as nn
-from torch.utils.data import DataLoader
 
 from nemo.collections.common.losses import AggregatorLoss, CrossEntropyLoss
 from nemo.collections.common.tokenizers.bert_tokenizer import NemoBertTokenizer
@@ -25,7 +23,6 @@ from nemo.collections.nlp.data.lm_bert_dataset import BertPretrainingPreprocesse
 from nemo.collections.nlp.losses import SmoothedCrossEntropyLoss
 from nemo.collections.nlp.modules.common import SequenceClassifier, TokenClassifier
 from nemo.collections.nlp.modules.common.common_utils import get_pretrained_lm_model
-from nemo.collections.nlp.modules.common.huggingface.bert import BertEncoder
 from nemo.core.classes import typecheck
 from nemo.core.classes.modelPT import ModelPT
 from nemo.core.neural_types import NeuralType
